messages: Initilization of left memebers
Fixes the coverity issues:

** 717282 Uninitialized scalar field
CID 717282 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member seq is not initialized
in this constructor nor in any functions that it calls.

** 717287 Uninitialized scalar field
CID 717287 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member tid is not initialized
in this constructor nor in any functions that it calls.

** 717288 Uninitialized scalar field
CID 717288 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member bits is not initialized
in this constructor nor in any functions that it calls.

** 717290 Uninitialized scalar field
CID 717290 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member epoch is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>