msg/msg_types: fix the dencoder of entity_addr_t
improve the interoperability between freebsd/osx and GNU/Linux, because
the their layouts of sockaddr_storage are different, and we use the
one of linux as the wire format. so need to convert it on freebsd/osx
side.

Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit 9099ca599de5238cde917f1e1f933247392de03e)