messages: Fixes uninitialized variables src/messages
Fixes the coverity issues:

** 717265 Uninitialized scalar field
2. uninit_member: Non-static class member field head.version is not initialized
in this constructor nor in any functions that it calls.
4. uninit_member: Non-static class member field head.oldest_client_tid is not
initialized in this constructor nor in any functions that it calls.
6. uninit_member: Non-static class member field head.mdsmap_epoch is not
initialized in this constructor nor in any functions that it calls.
8. uninit_member: Non-static class member field head.flags is not initialized
in this constructor nor in any functions that it calls.
10. uninit_member: Non-static class member field head.num_retry is not
initialized in this constructor nor in any functions that it calls.
12. uninit_member: Non-static class member field head.num_fwd is not
initialized in this constructor nor in any functions that it calls.
14. uninit_member: Non-static class member field head.num_releases is
not initialized in this constructor nor in any functions that it calls.
16. uninit_member: Non-static class member field head.op is not initialized
in this constructor nor in any functions that it calls.
18. uninit_member: Non-static class member field head.caller_uid is not
initialized in this constructor nor in any functions that it calls.
20. uninit_member: Non-static class member field head.caller_gid is not
initialized in this constructor nor in any functions that it calls.
CID 717265 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
22. uninit_member: Non-static class member field head.ino is not
initialized in this constructor nor in any functions that it calls.

** 717267 Uninitialized scalar field
2. uninit_member: Non-static class member field head.op is not
initialized in this constructor nor in any functions that it calls.
4. uninit_member: Non-static class member field head.seq is not
initialized in this constructor nor in any functions that it calls.
6. uninit_member: Non-static class member field head.stamp is not
initialized in this constructor nor in any functions that it calls.
8. uninit_member: Non-static class member field head.max_caps is
not initialized in this constructor nor in any functions that it calls.
CID 717267 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
10. uninit_member: Non-static class member field head.max_leases is
not initialized in this constructor nor in any functions that it calls.

** 717276 Uninitialized scalar field
CID 717276 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member success is not initialized
in this constructor nor in any functions that it calls.

** 717277 Uninitialized scalar field
CID 717277 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member ack is not initialized in
this constructor nor in any functions that it calls.

** 717292 Uninitialized scalar field
2. uninit_member: Non-static class member attempt is not initialized
in this constructor nor in any functions that it calls.
4. uninit_member: Non-static class member op is not initialized in
this constructor nor in any functions that it calls.
6. uninit_member: Non-static class member flags is not initialized
in this constructor nor in any functions that it calls.
8. uninit_member: Non-static class member lock_type is not initialized
in this constructor nor in any functions that it calls.
10. uninit_member: Non-static class member inode_export_v is not
initialized in this constructor nor in any functions that it calls.
CID 717292 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
12. uninit_member: Non-static class member srcdn_auth is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>