rbd: Initialization of m_state,tag_tid
Fixes the coverity issues:

** 1396177 Uninitialized scalar field
CID 1396177 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member m_state is not initialized
in this constructor nor in any functions that it calls.

** 1396178 Uninitialized scalar field
CID 1396178 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member tag_tid is not initialized
in this constructor nor in any functions that it calls.

** 1396203 Uninitialized scalar field
CID 1396203 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member m_state is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>