rgw: Initialization of sync_tracer,rgwx_stat,calc_md5
Fixes the coverity issues:

** 717382 Uninitialized pointer field
CID 717382 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
2. uninit_member: Non-static class member sync_tracer is not
initialized in this constructor nor in any functions that it calls.

** 717379 Uninitialized scalar field
CID 717379 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member rgwx_stat is not initialized
in this constructor nor in any functions that it calls.

** 1414873 Uninitialized scalar variable
CID 1414873 (#1 of 1): Uninitialized scalar variable (UNINIT)
25. uninit_use_in_call: Using uninitialized element of array
calc_md5 when calling strcmp.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>