client,osd,test: Initialize fuse_req_key,snap,who,seq
Fixes the coverity issues:

** 1351735 Uninitialized scalar field
CID 1351735 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member fuse_req_key is not
initialized in this constructor nor in any functions that it calls.

** 1396108 Uninitialized scalar field
CID 1396108 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member snap is not initialized
in this constructor nor in any functions that it calls.

** 1396115 Uninitialized scalar field
2. uninit_member: Non-static class member who is not initialized
in this constructor nor in any functions that it calls.
CID 1396115 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member seq is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>