os,librbd: Initialize gc_end,bits,m_object_exist
Fixes the coverity issues:

** 1401444 Uninitialized scalar field
2. uninit_member: Non-static class member gc_start_offset is not
initialized in this constructor nor in any functions that it calls.
CID 1401444 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member gc_end_offset is not
initialized in this constructor nor in any functions that it calls.

** 1405091 Uninitialized scalar field
CID 1405091 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member bits is not initialized
in this constructor nor in any functions that it calls.

** 1409842 Uninitialized scalar field
CID 1409842 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member m_object_exist is not
initialized in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>