msg: Initialize lkey,bound,port_cnt,num_chunk
Fixes the coverity issues:

** 1414512 Uninitialized scalar field
2. uninit_member: Non-static class member lkey is not initialized
in this constructor nor in any functions that it calls.
CID 1414512 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member bound is not initialized
in this constructor nor in any functions that it calls.

** 1414523 Uninitialized scalar field
CID 1414523 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
5. uninit_member: Non-static class member port_cnt is not initialized
in this constructor nor in any functions that it calls.

** 1414524 Uninitialized scalar field
CID 1414524 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member num_chunk is not initialized
in this constructor nor in any functions that it calls.

** 1414525 Uninitialized scalar field
CID 1414525 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member gid_idx is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>