osd,messages: Initialize read_length,options,send_reply
Fixes the coverity issues:

** 1415776 Uninitialized scalar field
CID 1415776 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member field read_length.v is
not initialized in this constructor nor in any functions that it calls.

** 1415811 Uninitialized scalar field
CID 1415811 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member options is not initialized
in this constructor nor in any functions that it calls.

** 1415850 Uninitialized scalar field
CID 1415850 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member sent_reply is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>