librbd: Initialize m_
Fixes the coverity issues:

** 1409838 Uninitialized scalar field
CID 1409838 (#1-2 of 2): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member m_watchers_ret_val is
not initialized in this constructor nor in any functions that it calls.

** 1405850 Uninitialized scalar field
2. uninit_member: Non-static class member reserv_heap_data is not
initialized in this constructor nor in any functions that it calls.
4. uninit_member: Non-static class member lim_heap_data is not
initialized in this constructor nor in any functions that it calls.
CID 1405850 (#1-3 of 3): Uninitialized scalar field (UNINIT_CTOR)
6. uninit_member: Non-static class member ready_heap_data is not
initialized in this constructor nor in any functions that it calls.

** 1413774 Uninitialized scalar field
CID 1413774 (#1-2 of 2): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member m_size is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>