tests: use special branch of ceph/s3-tests with pre-10.2.10
Jewel v10.2.10 introduces a fix for S3 ACL code, for which a new test was added
to ceph/s3-tests.git (ceph-jewel branch). Since the jewel point-to-point-x
upgrade test runs s3-tests on 10.2.7, modify the test to use a special
ceph/s3-tests branch (ceph-jewel-10-2-7) that omits the new test.

Signed-off-by: Nathan Cutler <ncutler@suse.com>