client: add delegation support for cephfs
Add the ability for ceph userland clients to request a delegation for a
cephfs open file. With this, userland will get a callback if there is a
request for access that conflicts with the delegation. The client is
then expected to return the delegation at which point the conflicting
access will succeed.

To support this, we also add a couple of configuration tunables:

Handing out a delegation means that we're trusting the application to
give back the caps when they are needed elsewhere. If it fails to
uphold its end of the bargain, we need to take steps to forcibly revoke
them, so that other clients can make progress.

When that occurs, shut down the mount such that further calls return
-ENOTCONN. That should prevent the application from holding on to caps
indefinitely, and ensure that it will no longer do damage. It should
also prevent the client from ending up blacklisted.

To support this, add a new config option (client_deleg_timeout) that
allows users and apps to tweak the timeout value. Note that this value
is not connected in any way to other ceph cluster cap timeouts, so
applications that tweak the delegation timeout should ensure that they
choose a value that prevents them from becoming blacklisted.

Note that we choose a default value of 75s to exceed ganesha's default
lease timeout.

An open call on ceph can easily return without any caps being granted to
the client. This is not generally a problem on ceph since it will wait
on caps before doing any operations on that open Fh, but it's less than
ideal when there is a delegation outstanding.

Both NFS and SMB servers can grant deny locks, and at least in the case
of NFS, those can be cached such that the server is unaware of them if
the client holds a delegation.

For correct delegation/oplock semantics we really want to break them on
a conflicting open call. Add a "client_deleg_break_on_open" boolean
tunable to allow the client to request that any delegations be broken
before returning from an open call.

Signed-off-by: Jeff Layton <jlayton@redhat.com>