mgr: safety checks on pyThreadState usage
Previously relied on the caller of Gil() to
pass new_thread=true if they would be
calling from a different thread.

Enforce this with an assertion, by wrapping
PyThreadState in a SafeThreadState class
that remembers which POSIX thread
it's meant to be used in.

Signed-off-by: John Spray <john.spray@redhat.com>
(cherry picked from commit 625e1b5cfb9b8a5843dfe75e97826f70a57d6ebe)