msgr: add a mechanism for Solaris to avoid dying on SIGPIPE
This is fairly clean: we define an RAII object in the Messenger.h on
Solaris, and "declare" it with a macro in the implementations. There's
no code duplication and on Linux it's just entirely compiled out.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit cba20a95e816aaf9f0971b29b14b0be8c524b59d)

Conflicts:
    src/msg/async/PosixStack.cc - luminous #include "common/simple_spin.h" is
        missing in master (trivial resolution)