osd/PGPool::update: optimize with deleting_snaps
Add a pg_pool_t.deleting_snaps member that tracks
the current set of snaps to be purged, in order to
avoid expensive operations involving the entire
set of removed snaps. The snaps are trimmed from
deleting snaps at the same time as they are added
to the pg_info_t.purged_snaps set.

Suggested-by: Greg Farnum <gfarnum@redhat.com>
Signed-off-by: Zac Medico <zmedico@gmail.com>