client: rip out return InodeRef pointer handling from setattr codepaths
We've had a report of the assertion in ll_setattr and ll_setattrx
being tripped [1]. Evidently we made a call to the server and that got
us back a trace with a different inode than the one we passed down.

In practice, nothing uses the pointer we're passing in to _do_setattr,
save the assertions, which I think are actually wrong. There's nothing
that prevents a concurrent rename or something from another client
while this setattr is in flight.

Since these assertions are questionable anyway, let's just remove them.

[1]: https://github.com/nfs-ganesha/nfs-ganesha/issues/215

Tracker: http://tracker.ceph.com/issues/21748
Signed-off-by: Jeff Layton <jlayton@redhat.com>