rgw: allow buckets created with / to be useable
Previously we didn't restrict "/" in bucket names created via swift api
though swift client itself had a restriction, and cases like
misconfigured keystone urls would cause swift client also to prepend a
"/" before the bucket name. The problem is that create bucket succeeds
such a call and rgw_get_bucket_instance_info() and related functions
would fail because we replace the first encountered "/" with a ":"
assuming that this was a(n empty) tenant name. Make these buckets
useable/ deleteable by making sure that when we retreive the key back we
replace the beginning / with a colon as this was used in storing the
bucket info. A non backwards compatible change would be to make
rgw_bucket_instance_key_to_oid not replace slashes when 0 sized tenants
were encountered. Swift clients are still broken when uploading objects
etc. as they do not prepend the / in PUT object requests whereas do in
create bucket requests. However this patch makes it possible to
use/delete these created buckets if using a rest api.

Note that however there is still an edge case which cannot be addressed,
which is, currently an object upload request of the form PUT /bucket/foo
would end up with another bucket with the name /bucket/foo being
created. With this patch however we ensure that a PUT /bucket/foo is
indeed considered as an object request, however this introduces the
problem where previously a bucket was created with such a name which
would no longer be accessible. There is no way to address this since
object names are allowed to contain slashes and we are now just allowing
the case where a prefix slash at bucket name is allowed (since we hadn't
restricted that before), a slash in the middle would end up with the
part after being considered as an object request, which is the convention.

Fixes: http://tracker.ceph.com/issues/21679
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>