msg/async/rdma: fix the deletion of an item of the dead_queue_pairs vector
by using 'dead_queue_pairs.erase(it)', we modify the vector and the
structure of the vector becomes inconsistent since the method returns
an iterator to the new position that has to be accessed in the next
iteration and  a for range-based loop does not allow updates which
would reflect the change. Due to this, it is possible to skip a few
members of the vector as the underlying iterator of the for range-based
loop is corrupt.

Some checks with stakcoverflow (for example, question: 10360461) explain
the bug and suggest using std::remove_if instead.

Signed-off-by: Justinas Lingys <jlingys@connect.ust.hk>