osd/PrimaryLogPG: primary_error - grab peers which have the right object version
The original API means either the object passed in is not missing
on that peer, or it needs a newer version. In the second case, we
might include a bad peer which still does not have the right version
we want to pull later.

Fix the above problem by comparing with what the peers have now
instead of what they need now.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>