doc: Remove additional arguments when replacing OSD
It's not required to provide the --bluestore nor --osd-uuid option
to ceph-disk.

Using BlueStore is the default on master (e.g. Luminous) and thus
this option is not required.

When no OSD UUID is supplied ceph-disk will generate one, so we do
not need to supply one generated by uuidgen.

Signed-off-by: Wido den Hollander <wido@42on.com>