os/bluestore: onode_map differentiates hashes of ghobjects.
Some time ago Adam Kupczyk (@aclamk) has discovered that
hashes of ghobjects constructed by the Ceph's FIO plugin
don't depend on the object's name but instead use the PG-
related data. This approach is conflicted with employing
the std::unordered_map as a BlueStore's Onode Map due to
increased size of its single bucket.

Unfortunately, Adam's fix caused a regression in get_onode()
and has to be reverted.

This patch tries to deal with the effect by providing BS'
onode_map with a hash function that considers also names.
The new global, approach could be beneficial as ghobjects
in other places seem to be constructed similarly (pg::m_seed
used as a hash).

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>